requirements.txt
# Core Dependencies
streamlit>=1.30.0
pypdf>=3.17.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
ollama>=0.1.0

# Configuration & Environment
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
